cmd/wit-bindgen-go/cmd/generate: directly pipe the stdin to wasm-tools and rename wkg to oci
